first stab at re-factoring fusion ring code. doc tests pass except for a couple issues with get_order and FusionRingFromWCR.__classcall__. Need some documentation/explanation in FusionRing constructor
